# src/auth/dto/send-invite.dto.ts
import { IsEmail, IsInt } from 'class-validator';
import { ApiProperty } from '@nestjs/swagger';

export class SendInviteDto {
  @ApiProperty({ example: 'user@example.com', description: 'User email address' })
  @IsEmail()
  email: string;

  @ApiProperty({ example: 1, description: 'Tenant ID' })
  @IsInt()
  tenantId: number;
}